Making ReceiptsView primary screen (Making title bigger and removing back button) As per the discussion with madhu, he expects the following 	1. User is in receipt details screen 	2. Taps Home tab in bottom nav tab bar. 	3. Then taps More tab, and see the more screen instead of where he left in receipts flow before. As a default behaviour, iOS shows where the user left in receipts flow when they tap on more.
In order to achieve this behaviour, we need to find a way to attach an action on selectedTab change in BottonNavTabView. it is done using onChange modifier.  In order to dismiss all the screens in the navigation stack, I have used path destination of NavigationStack. so it is easy to remove all screens from the path.