Issue: When user goes to voucher section from congratsview (by tapping played game in gamezone) and comes back to game zone by tapping back button, and when they tap on active tab, it tab changes but the view remains in played games view.
Fix: Add tabSelection property in GameZoneViewModel and pass it as binding to GameZoneTabView. And have a state property in Gamezoneview to pass it as binding to TopTabBar. And have logic to sync these two properties. Ideally, there should be one source of truth. But in respect of time, I'm just keeping this as a temporary fix and put this issue in backlog.